implemented upload to currentDir feature
