+++ Server/source/index.ts
@@ -1,27 +1,14 @@
 import * as express from 'express';
 import * as cors from 'cors';
-import * as multer from "multer";
-import * as path from "path";
+
 
 import { IRequest, IResponse } from './interfaces/express.interface';
 import { getFileFromPath } from './modules/getFileFromPath';
 import { getListedFilesWithInfo } from "./modules/getListedFilesWithInfo";
 import { postFile } from './modules/postFile';
-import { IMulterFile, IMulterDestinationCallback, IMulterFileNameCallback, IMulter } from './interfaces/multer.interface';
 import { createDirectory } from './modules/createDirectory';
 import { deleteDirectory } from './modules/deleteDirectory';
 
-const fileStorage = multer.diskStorage({
-    destination: (_request: IRequest, file: IMulterFile, callback: IMulterDestinationCallback) => {
-        console.log(`Uploading file: ${file.originalname}`);
-        callback(null, path.join(__dirname, "../Database"));
-    },
-    filename: (_request: IRequest, file: IMulterFile, callback: IMulterFileNameCallback) => {
-        callback(null, file.originalname);
-    }
-});
-
-const upload: IMulter = multer({ storage: fileStorage });
 
 const app = express();
 const { PORT = 3001} = process.env;
@@ -29,7 +16,7 @@ app.use(cors());
 app.use(express.json());
 getListedFilesWithInfo(app);
 getFileFromPath(app);
-postFile(app, upload);
+postFile(app);
 createDirectory(app);
 deleteDirectory(app);
