+++ Server/source/modules/postFile.ts
@@ -1,8 +1,25 @@
 import { IApp, IRequest, IResponse } from "../interfaces/express.interface";
-import { IMulter } from "../interfaces/multer.interface";
+import * as multer from "multer";
+import * as path from "path";
+import { IMulterFile, IMulterDestinationCallback, IMulterFileNameCallback, IMulter } from "../interfaces/multer.interface";
 
-export const postFile = (app: IApp, upload: IMulter) => {
+
+const fileStorage = multer.diskStorage({
+    destination: (_request: IRequest, file: IMulterFile, callback: IMulterDestinationCallback) => {
+        const { filePath } =_request.body;
+        console.log(`Uploading file: ${file.originalname}`);
+        callback(null, filePath);
+    },
+    filename: (_request: IRequest, file: IMulterFile, callback: IMulterFileNameCallback) => {
+        callback(null, file.originalname);
+    }
+});
+
+const upload: IMulter = multer({ storage: fileStorage });
+
+export const postFile = (app: IApp) => {
     app.post('/uploadFile', upload.any(), (_request: IRequest, response: IResponse) => {
         response.sendStatus(200);
     });
 }
+
